requirements.txt
langchain
langchain-ollama
langchain-community
pydantic
unstructured
faiss-cpu
uvicorn
gunicorn
python-dotenv
autopep8